can send invite, cant accept/reject

+++ app/Http/Controllers/InvitedController.php
@@ -17,16 +17,24 @@ class InvitedController extends Controller
     /**
      * Accepts an invite by updating accept boolean of invite to true
      */
-    public function accept($invited_user_id, $inviter_user_id, $event_id){
-        $invite = Invited::find([$invited_user_id, $inviter_user_id,$event_id]);
-        $this->authorize('join', $event);
+    public function accept(Request $request){
+        if (!Auth::check()) {return response(route('login'), 302);}
+
+        $invited_user_id = Auth::user()->userid;
+        $event_id = $request['event_id'];
+
+        $invite = Invited::where('invited.inviteduserid', '=',  $invited_user_id)
+                        ->where('invited.eventid', '=', $event_id)
+                        ->first();
+
         $this->authorize('update', $invite);
+        $this->authorize('join', $event);
 
         $invite->status = TRUE;
         $invite->save();
         $event = $invite->event()->first();
         
-        app('App\Http\Controllers\EventController')->join($invite->eventID);
+        app('App\Http\Controllers\EventController')->join($invite->eventid);
 
         return response(null, 200);
     }
@@ -34,31 +42,49 @@ class InvitedController extends Controller
     /**
     * Rejects an invite by removing it from the database
     */
-    public function reject($invited_user_id, $inviter_user_id,$event_id) {
-        $invite = Invite::find([$invited_user_id, $inviter_user_id,$event_id]);
+    public function reject(Request $request) {
+        if (!Auth::check()) {return response(route('login'), 302);}
+
+        $invited_user_id = Auth::user()->userid;
+        $event_id = $request['event_id'];
+
+        $invite = Invited::where('invited.inviteduserid', '=',  $invited_user_id)
+                        ->where('invited.eventid', '=', $event_id)
+                        ->first();
+        
+        if(!$invite){
+            return response("invite doesnt exist", 313);
+        }
+        
         $this->authorize('delete', $invite);
+        return $invite;
         $invite->delete();
+        
         return response(null, 200);
     }
 
     public function create(InviteRequest $request) {
-        return response($request, 201);
-        $this->validated();
-        if (!Auth::check()) return redirect('/login');
+        //$this->validate($request); 
+        if (!Auth::check()) {return response(route('login'), 302);}
 
         $invited_user_id = $request['invited_user'];
-        $inviter_user_id = Auth::user()->userID;
+        $inviter_user_id = Auth::user()->userid;
         $event_id = $request['event_id'];
 
-        $invite = Invite::findOr([$invited_user_id, $inviter_user_id,$event_id], function($invited_user_id, $inviter_user_id,$event_id) {
+        $invite = Invited::where('invited.inviteduserid', '=',  $invited_user_id)
+                        ->where('invited.inviteruserid', '=', $inviter_user_id)
+                        ->where('invited.eventid', '=', $event_id)
+                        ->first();
+        if(!$invite){
             $inv = new Invited;
-            $inv->invitedUserid = $invited_user_id;
-            $inv->inviterUserID = $inviter_user_id;
+            $inv->inviteduserid = $invited_user_id;
+            $inv->inviteruserid = $inviter_user_id;
             $inv->eventid = $event_id;
             $inv->save();
-        });
+            return response(null, 200);
+        }
 
-        return response($request, 200);
+        return response(null, 409);
     }
 
 }
